Rewrite brew.fish to use one `for loop` only
With this change, a path is added to `$PATH` only if it is not already
part of `$PATH` and contains a file called `brew`. Also, the path is
added in the beginning of `$PATH` to ensure it has precedence over
system paths, otherwise system apps like `ruby` and `python` would be
preferred over the Homebrew variants. It should be the other way around.